Phase_type

Description of the S3 classes cont_phase_type, disc_phase_type, mult_cont_phase_type, mult_disc_phase_type which represents the different phase-type distributions.

phase_type(subint_mat = NULL, init_probs = NULL, reward_mat = NULL,
                  round_zero = NULL)

# S3 method for cont_phase_type
summary(object, ...)

Arguments

subint_mat
is the square matrix containing the transition rates or probabilities between transient states for continous or discrete phase-type respectively. If the phase-type is continuous, the subintensity matrix diagonal should only contains negative values and the row sums should be non-positive. If the phase-type is discrete, the subintensity matrix should only contains values between 0 and 1.

init_probs
a vector, a one-row matrix or NULL which gives the probabilities to start in each state. If init_probs is NULL, the probability to start on the first state will be 1 and to start on any other state 0.

reward_mat
is a matrix codeNULL(default) where each row is a reward vector, and each column corresponds to a state. It should have the same number of columns as the length of the initial probabilities.

round_zero
is an integer or NULL(default), which gives the decimal from which we should truncate the positive values of the subintensity matrix. It could be useful in the scenarios where there is a reward transformation leading to values with many decimals and potentially computational approximation and potentially to positive row sums in continuous phase-type .

object
a phase-type object

...
other arguments passed to methods

Details

phase_type is the generator function for the four types of phase-type classes, respectively univariate continuous or discrete and multivariate continuous or discrete which inherits from list. The class is generated by supplying a sub-intensity matrix and an optional initial probability vector plus a reward matrix in the case of multivariate phase-type. If the initial probabilities are not specified, then the initial probability will be init_probs = c(1, 0, 0, ...) with the same length as the number of transient states.

Methods (by generic)

summary: pretty summary of the class.

Examples

##===========================##
## For continuous univariate ##
##===========================##

subintensity_matrix <- matrix(c(-1.5, 0, 0,
                               1.5, -1, 0,
                               0, 1, -0.5), ncol = 3)
phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,] -1.5  1.5  0.0
#> [2,]  0.0 -1.0  1.0
#> [3,]  0.0  0.0 -0.5
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "cont_phase_type"

#---

subintensity_matrix <- matrix(c(-1.5, 0, 0,
                               1.5, -1, 0,
                               0, 1, -0.5), ncol = 3)
initial_probabilities <- c(0.9, 0.1, 0)
phase_type(subintensity_matrix, initial_probabilities)

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,] -1.5  1.5  0.0
#> [2,]  0.0 -1.0  1.0
#> [3,]  0.0  0.0 -0.5
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]  0.9  0.1    0
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "cont_phase_type"

##=========================##
## For discrete univariate ##
##=========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]  0.4 0.24 0.12
#> [2,]  0.0 0.40 0.20
#> [3,]  0.0 0.00 0.50
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "disc_phase_type"

#---

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
initial_probabilities <- c(0.9, 0.1, 0)
phase_type(subintensity_matrix, initial_probabilities)

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]  0.4 0.24 0.12
#> [2,]  0.0 0.40 0.20
#> [3,]  0.0 0.00 0.50
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]  0.9  0.1    0
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "disc_phase_type"

##=============================##
## For continuous multivariate ##
##=============================##

subintensity_matrix <- matrix(c(-3, 0, 0,
                               2, -2, 0,
                               0, 1, -1), nrow = 3, ncol = 3)
reward_matrix = matrix(sample(seq(0, 10, 0.1), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]   -3    2    0
#> [2,]    0   -2    1
#> [3,]    0    0   -1
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $reward_mat
#>      [,1] [,2]
#> [1,]  1.6  4.7
#> [2,]  7.3  3.8
#> [3,]  1.2  8.4
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "mult_cont_phase_type"

##===========================##
## For discrete multivariate ##
##===========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
reward_matrix <- matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

#> Warning: the condition has length > 1 and only the first element will be used

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]  0.4 0.24 0.12
#> [2,]  0.0 0.40 0.20
#> [3,]  0.0 0.00 0.50
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $reward_mat
#>      [,1] [,2]
#> [1,]    3    6
#> [2,]    0    9
#> [3,]    2    5
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "mult_disc_phase_type"