Phase-type simulations

sim_phase_type implements a naive simulation scheme for simulation of phase-type distributions.

sim_phase_type implements a naive simulation scheme for simulation of phase-type distributions.

sim_phase_type <- function(obj)

sim_phase_type <- function(obj)

Arguments

obj
an object of class disc_phase_type or cont_phase_type

Examples

##===========================##
## For continuous univariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(-1.5, 0, 0.5,
                               1.5, -1, 0,
                               0, 1, -1), ncol = 3)
obj <- phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

sim_phase_type(obj)

#> [[1]]
#>  [1] 0.00000000 0.26983205 1.73070972 0.08952618 0.06664499 1.51162930
#>  [7] 0.15653028 1.51940822 1.24334535 1.59030589 0.53106482 4.49194233
#> [13] 0.62038429
#> 
#> [[2]]
#>  [1] 1 2 3 1 2 3 1 2 3 1 2 3
#> 

##===========================##
## For discrete multivariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(0.4, 0, 0.1,
                               0.24, 0.6, 0.3,
                               0.26, 0.4, 0.5), ncol = 3)
obj <- phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

sim_phase_type(obj)

#> [[1]]
#> [1] 1 3 1
#> 
#> [[2]]
#> [1] 1 2 3
#> 

##=============================##
## For continuous multivariate ##
##=============================##

subintensity_matrix <- matrix(c(-3, 0, 0,
                               2, -2, 0,
                               0, 1, -1), nrow = 3, ncol = 3)
reward_matrix = matrix(sample(seq(0, 10, 0.1), 6), nrow = 3, ncol = 2)
obj <- phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

sim_phase_type(obj)

#> Error in sim_phase_type(obj): Please provide an object of 'class disc_phase_type' or
#>          'cont_phase_type'

##===========================##
## For discrete multivariate ##
##===========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
reward_matrix <- matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]  0.4 0.24 0.12
#> [2,]  0.0 0.40 0.20
#> [3,]  0.0 0.00 0.50
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $reward_mat
#>      [,1] [,2]
#> [1,]   10    5
#> [2,]    0    1
#> [3,]    2    8
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "mult_disc_phase_type"

##===========================##
## For continuous univariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(-1.5, 0, 0.5,
                               1.5, -1, 0,
                               0, 1, -1), ncol = 3)
obj <- phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

sim_phase_type(obj)

#> [[1]]
#>  [1] 0.00000000 0.26983205 1.73070972 0.08952618 0.06664499 1.51162930
#>  [7] 0.15653028 1.51940822 1.24334535 1.59030589 0.53106482 4.49194233
#> [13] 0.62038429
#> 
#> [[2]]
#>  [1] 1 2 3 1 2 3 1 2 3 1 2 3
#> 

##===========================##
## For discrete multivariate ##
##===========================##

set.seed(2)
subintensity_matrix <- matrix(c(0.4, 0, 0.1,
                               0.24, 0.6, 0.3,
                               0.26, 0.4, 0.5), ncol = 3)
obj <- phase_type(subintensity_matrix)

#> Warning: The initial probability vector is automatically generated.

sim_phase_type(obj)

#> [[1]]
#> [1] 1 3 1
#> 
#> [[2]]
#> [1] 1 2 3
#> 

##=============================##
## For continuous multivariate ##
##=============================##

subintensity_matrix <- matrix(c(-3, 0, 0,
                               2, -2, 0,
                               0, 1, -1), nrow = 3, ncol = 3)
reward_matrix = matrix(sample(seq(0, 10, 0.1), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]   -3    2    0
#> [2,]    0   -2    1
#> [3,]    0    0   -1
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $reward_mat
#>      [,1] [,2]
#> [1,]  1.6  7.5
#> [2,]  9.2  4.0
#> [3,]  8.0  4.9
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "mult_cont_phase_type"

##===========================##
## For discrete multivariate ##
##===========================##

subintensity_matrix <- matrix(c(0.4, 0, 0,
                               0.24, 0.4, 0,
                               0.12, 0.2, 0.5), ncol = 3)
reward_matrix <- matrix(sample(seq(0, 10), 6), nrow = 3, ncol = 2)
phase_type(subintensity_matrix, reward_mat = reward_matrix)

#> Warning: The initial probability vector is automatically generated.

#> $subint_mat
#>      [,1] [,2] [,3]
#> [1,]  0.4 0.24 0.12
#> [2,]  0.0 0.40 0.20
#> [3,]  0.0 0.00 0.50
#> 
#> $init_probs
#>      [,1] [,2] [,3]
#> [1,]    1    0    0
#> 
#> $reward_mat
#>      [,1] [,2]
#> [1,]   10    5
#> [2,]    0    1
#> [3,]    2    8
#> 
#> $defect
#> [1] 0
#> 
#> attr(,"class")
#> [1] "mult_disc_phase_type"